let the pgo task fail again
test_os often fails for me when run inside the debian build system

ubuntu lets the pgo task fail, but I had turned that off because I kept breaking the pgo task and we ended up with pgo-less builds. that hasn't happened in a while so I prefer to let it fail again.