make_order.py: traverse through noarch packages
Previously I stopped the traversal at any package we don't need
to build, but then I ran into a case of a noarch package that depends
on a binary package, ie we have to build that transitive dependency.

So change the logic so that we continue the traversal for these
packages even though we don't output them for building.