Add call_function "aot trace" to pyston_lite
Doesn't use any of the actual "aot" machinery; I just copied in
all of the relevant functions and specialized them by hand, and the
inliner takes care of merging them into a single function.

This commit only adds specializations for calling Python functions / methods.
It is not specialized for the argument count, unlike pyston-full.

About a 2% speedup on pyperformance and 1% on macrobenchmarks