pyston_lite: build linux packages for all supported CPython versions Adopts 'make package' to build wheels for all our supported versions.
We use the the manylinux docker image and build against installed python versions it distributes but
for BOLT we have to run it outside the host so we will need to have the different python versions installed.
I found the easiest way is by using the deadsnakes PPA.

This also showed up a problem:
The manylinux CPython has different defines set which run into a case which I have not had ported.
In general this is quite ugly I wish they would just export '_PyThread_cond_after' / 'PyCOND_TIMEDWAIT' so that we
don't have to copy this in.